Converted the Sobol' rotation matrices to float32 in the frame order target function.
This is to conserve huge amounts of memory to allow for more Sobol' points to be used.  For example
for the models which use 3D Sobol' points (isotropic cone and pseudo-ellipse), a maximum of 50000
Sobol' points requires 50000000 to be created, using about 15 Gb of RAM.

+++ target_functions/frame_order.py
@@ -2031,8 +2031,8 @@ class Frame_order:
         sobol_data.model = self.model
         sobol_data.total_num = total_num
         sobol_data.sobol_angles = zeros((m, total_num), float32)
-        sobol_data.Ri_prime = zeros((total_num, 3, 3), float64)
-        sobol_data.Ri2_prime = zeros((total_num, 3, 3), float64)
+        sobol_data.Ri_prime = zeros((total_num, 3, 3), float32)
+        sobol_data.Ri2_prime = zeros((total_num, 3, 3), float32)
 
         # The Sobol' points.
         points = i4_sobol_generate(m, total_num, 1000)
